Add: Parsing with `std::regex` & `ctre::`
This commit adds a simple config parser
using RegEx engines - the one native to
STL, and the modern expression templates
CTRE library.

CTRE provides excellent performance,
surpassing every approach except SZ
on short configs, and losing only to SZ
and STL parser on longer inputs.

Compare to `std::regex` with a pre-
constructed automata it is 10x faster.

Important to note, there is an issue with
handling the `$` end-of-line anchor in
CTRE:
https://github.com/hanickadot/compile-time-regular-expressions/pull/333